requirements_dev.txt
pip==20.2.1
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==8.0
requests~=2.25.1
rich~=10.1.0
proxmoxer~=1.1.1
setuptools~=49.2.1
PyYAML==5.4.1
defusedxml~=0.7.1
requests-oauthlib==1.3.0
botocore==1.21.1
boto3==1.18.1
